requirements.txt
requests==2.32.3
espn-api>=0.44.1
pandas==2.2.3
datetime==5.5
sqlalchemy==2.0.37
python-dotenv==1.0.1
psycopg2-binary==2.9.10